# doc/api/swagger.yaml
swagger: '2.0'
info:
  title: Fractalis API
  description: foobar
  version: 0.0.0
host: '127.0.0.1:5000'
schemes:
  - http
produces:
  - application/json
paths:
  /data:
    post:
      summary: Add data to session
      parameters:
        - name: handler
          in: query
          description: Handler to use for moving data from source to Fractalis
          required: true
          type: string
        - name: server
          in: query
          description: Server that is used as a data source
          required: true
          type: string
        - name: auth
          in: query
          description: Contains necessary information to authenticate with server API
          required: true
          type: object
        - name: descriptors
          in: query
          description: >-
            Contains information necessary to request certain data from the
            server
          required: true
          type: object
      responses:
        '201':
          description: CREATED
          schema:
            type: object
            properties:
              data_ids:
                type: array
                items:
                  type: string
    get:
      summary: Get all completed and running data jobs for session
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              data_states:
                type: array
                items:
                  $ref: '#/definitions/DataState'
  '/data/{data_id}':
    parameters:
      - name: data_id
        in: path
        description: ID given on launching a data job
        required: true
        type: string
    get:
      summary: Get status of data job associated with data_id
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              data_state:
                $ref: '#/definitions/DataState'
  /analytics:
    post:
      summary: Submit analysis job
      parameters:
        - name: job_name
          in: query
          description: Analysis job to run
          required: true
          type: string
        - name: args
          in: query
          description: Arguments passed to the analysis job
          required: true
          type: object
      responses:
        '201':
          description: CREATED
          schema:
            type: object
            properties:
              job_id:
                type: string
  '/analytics/{job_id}':
    parameters:
      - name: job_id
        in: path
        description: ID given on job creation
        required: true
        type: string
    get:
      summary: Get job status information for given job id
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/JobState'
    delete:
      summary: Attempts to kill all processes associated with the job
      responses:
        '200':
          description: OK
definitions:
  DataState:
    type: object
    properties:
      data_id:
        type: string
      data_type:
        type: string
      description:
        type: string
      job_id:
        type: string
      message:
        type: string
      state:
        type: string
  JobState:
    type: object
    properties:
      state:
        type: string
      result:
        type: string